api/requirements.txt
uvicorn~=0.15.0
fastapi~=0.70.0
pydantic~=1.8.2
starlette~=0.16.0
PyYAML~=6.0
pytz~=2021.3
asyncpg~=0.24.0